refactor code - multiple files
